refactor(turbo-tasks-fs): move file watching into separate file
