+++ Cargo.toml
@@ -7,7 +7,7 @@ members = [
   "crates/turbo-tasks*",
   "crates/turbopack*",
   "crates/turborepo*",
-  "packages/turbo-repository/rust/",
+  "packages/turbo-repository/rust",
   "xtask",
 ]
 
@@ -230,6 +230,9 @@ lightningcss = { version = "1.0.0-alpha.50", features = [
 ] }
 mime = "0.3.16"
 nohash-hasher = "0.2.0"
+notify = "6.1.1"
+notify-debouncer-full = "0.3.1"
+notify-debouncer-mini = { version = "0.3.0", default-features = false }
 once_cell = "1.17.1"
 owo-colors = "3.5.0"
 parcel_selectors = "0.26.0"
@@ -278,4 +281,3 @@ urlencoding = "2.1.2"
 webbrowser = "0.8.7"
 which = "4.4.0"
 unicode-segmentation = "1.10.1"
-notify-debouncer-full = "0.3.1"
