+++ crates/turborepo-filewatch/Cargo.toml
@@ -14,7 +14,7 @@ anyhow = { workspace = true }
 dashmap = { workspace = true }
 futures = { version = "0.3.26" }
 itertools = { workspace = true }
-notify = "6.0.1"
+notify = { workspace = true }
 thiserror = "1.0.38"
 tokio = { workspace = true, features = ["full", "time"] }
 tracing = "0.1.37"
