+++ crates/turborepo-lib/Cargo.toml
@@ -62,7 +62,7 @@ jsonc-parser = { version = "0.21.0" }
 lazy_static = { workspace = true }
 libc = "0.2.140"
 nix = "0.26.2"
-notify = "5.1"
+notify = { workspace = true }
 path-clean = "1.0.1"
 petgraph = { workspace = true }
 pidlock = { path = "../turborepo-pidlock" }
